Module aws_lc_rs::aead::nonce_sequence

Implementations of NonceSequence for use with BoundKeys.

Structs

Counter32

Counter32 is an implementation of the NonceSequence trait. The internal state of a Counter32 is a 32-bit unsigned counter that increments on each call to advance and an optional 8-byte identifier. Counter and identifier values are used to construct each nonce. A limit can be set on the number of nonces allowed to be generated, by default this limit is u32::MAX.

Counter32Builder

NonceSequenceBuilder facilitates the building of a Counter32.

Counter64

Counter64 is an implementation of the NonceSequence trait. The internal state of a Counter64 is a 64-bit unsigned counter that increments on each call to advance and an optional 4-byte identifier. Counter and identifier values are used to construct each nonce. A limit can be set on the number of nonces allowed to be generated, by default this limit is u64::MAX. See Section 3.2 of RFC 5116.

Counter64Builder

NonceSequenceBuilder facilitates the building of a Counter64.