Struct aws_lc_rs::aead::nonce_sequence::Counter32

pub struct Counter32 { /* private fields */ }

Counter32 is an implementation of the NonceSequence trait. The internal state of a Counter32 is a 32-bit unsigned counter that increments on each call to advance and an optional 8-byte identifier. Counter and identifier values are used to construct each nonce. A limit can be set on the number of nonces allowed to be generated, by default this limit is u32::MAX.

See Section 3.2 of RFC 5116.

Implementations

impl Counter32

pub fn identifier(&self) -> [u8; 8]

Provides the internal identifier.

pub fn counter(&self) -> u32

Provides the current internal counter value.

pub fn generated(&self) -> u32

Provides the current counter indicating how many nonces have been generated.

pub fn limit(&self) -> u32

Provides the limit on the number of nonces allowed to be generate.