aws_lc_rs::aead
Struct OpeningKeyPreparedNonce

pub struct OpeningKeyPreparedNonce<'a, N: NonceSequence> { /* private fields */ }

A key operation with a precomputed nonce from a key’s associated NonceSequence.

Implementations

impl<N: NonceSequence> OpeningKeyPreparedNonce<'_, N>

pub fn nonce(&self) -> &Nonce

Returns the prepared Nonce that is used for key methods invoked on Self.

impl<N: NonceSequence> OpeningKeyPreparedNonce<'_, N>

pub fn open_in_place<A>( self, aad: Aad<A>, in_out: &mut [u8], ) -> Result<&mut [u8], Unspecified>
where A: AsRef<[u8]>,

Authenticates and decrypts (“opens”) data in place.

See OpeningKey::open_in_place for additional API information.

Errors

error::Unspecified when ciphertext is invalid. In this case, in_out may have been overwritten in an unspecified way.

pub fn open_within<A>( self, aad: Aad<A>, in_out: &mut [u8], ciphertext_and_tag: RangeFrom<usize>, ) -> Result<&mut [u8], Unspecified>
where A: AsRef<[u8]>,

Authenticates and decrypts (“opens”) data in place, with a shift.

See OpeningKey::open_within for additional API information.

Errors

error::Unspecified when ciphertext is invalid. In this case, in_out may have been overwritten in an unspecified way.