Struct aws_lc_rs::agreement::PrivateKey

pub struct PrivateKey { /* private fields */ }

A private key for use (only) with agree. The signature of agree allows PrivateKey to be used for more than one key agreement.

Implementations

impl PrivateKey

pub fn generate(alg: &'static Algorithm) -> Result<Self, Unspecified>

Generate a new private key for the given algorithm.

Errors

error::Unspecified when operation fails due to internal error.

pub fn from_private_key_der( alg: &'static Algorithm, key_bytes: &[u8] ) -> Result<Self, KeyRejected>

Deserializes a DER-encoded private key structure to produce a agreement::PrivateKey.

This function is typically used to deserialize RFC 5915 encoded private keys, but it will attempt to automatically detect other key formats. This function supports unencrypted PKCS#8 PrivateKeyInfo structures as well as key type specific formats.

X25519 keys are not supported. See PrivateKey::as_der.

Errors

error::KeyRejected if parsing failed or key otherwise unacceptable.

Panics

pub fn from_private_key( alg: &'static Algorithm, key_bytes: &[u8] ) -> Result<Self, KeyRejected>

Constructs an ECDH key from private key bytes

The private key must encoded as a big-endian fixed-length integer. For example, a P-256 private key must be 32 bytes prefixed with leading zeros as needed.

Errors

error::KeyRejected if parsing failed or key otherwise unacceptable.

pub fn compute_public_key(&self) -> Result<PublicKey, Unspecified>

Computes the public key from the private key.

Errors

error::Unspecified when operation fails due to internal error.

pub fn algorithm(&self) -> &'static Algorithm

The algorithm for the private key.