aws_lc_rs::cipher

Struct StreamingEncryptingKey

pub struct StreamingEncryptingKey { /* private fields */ }

A key for streaming encryption operations.

Implementations

impl StreamingEncryptingKey

pub fn update<'a>( &mut self, input: &[u8], output: &'a mut [u8], ) -> Result<BufferUpdate<'a>, Unspecified>

Updates the internal state of the key with the provided ciphertext input, potentially writing bytes of ciphertext to output.

The number of bytes written to output can be up to input.len() plus the block length of the algorithm (e.g., Algorithm::block_len).

Errors

Returns an error if the output buffer is smaller than the length of the input plus the algorithm’s block length (e.g. Algorithm::block_len) minus one.

May return an error if the length of input plus the algorithm’s block length is larger than i32::MAX.

pub fn finish( self, output: &mut [u8], ) -> Result<(DecryptionContext, BufferUpdate<'_>), Unspecified>

Finishes the encryption operation, writing any remaining ciphertext to output.

The number of bytes written to output can be up to the block length of Algorithm::block_len.

Errors

Returns an error if the output buffer is smaller than the algorithm’s block length.

pub fn mode(&self) -> OperatingMode

Returns the cipher operating mode.

pub fn algorithm(&self) -> &'static Algorithm

Returns the cipher algorithm.

pub fn ctr(key: UnboundCipherKey) -> Result<Self, Unspecified>

Constructs a StreamingEncryptingKey for encrypting data using the CTR cipher mode. The resulting ciphertext will be the same length as the plaintext.

Errors

Returns and error on an internal failure.

pub fn less_safe_ctr( key: UnboundCipherKey, context: EncryptionContext, ) -> Result<Self, Unspecified>

Constructs a StreamingEncryptingKey for encrypting data using the CTR cipher mode. The resulting ciphertext will be the same length as the plaintext.

This is considered less safe because the caller could potentially construct an EncryptionContext from a previously used initialization vector (IV).

Errors

Returns an error on an internal failure.

pub fn cbc_pkcs7(key: UnboundCipherKey) -> Result<Self, Unspecified>

Constructs a StreamingEncryptingKey for encrypting data using the CBC cipher mode with pkcs7 padding. The resulting ciphertext will be longer than the plaintext; padding is added to fill the next block of ciphertext.

Errors

Returns an error on an internal failure.

pub fn cfb128(key: UnboundCipherKey) -> Result<Self, Unspecified>

Constructs a StreamingEncryptingKey for encrypting data using the CFB128 cipher mode. The resulting ciphertext will be the same length as the plaintext.

Errors

Returns and error on an internal failure.

pub fn ecb_pkcs7(key: UnboundCipherKey) -> Result<Self, Unspecified>

Constructs a StreamingEncryptingKey for encrypting using ECB cipher mode with PKCS7 padding. The resulting plaintext will be the same length as the ciphertext.

☠️ ️️️DANGER ☠️

Offered for computability purposes only. This is an extremely dangerous mode, and very likely not what you want to use.

Errors

Returns an error on an internal failure.

pub fn less_safe_cfb128( key: UnboundCipherKey, context: EncryptionContext, ) -> Result<Self, Unspecified>

Constructs a StreamingEncryptingKey for encrypting data using the CFB128 cipher mode. The resulting ciphertext will be the same length as the plaintext.

This is considered less safe because the caller could potentially construct an EncryptionContext from a previously used initialization vector (IV).

Errors

Returns an error on an internal failure.

pub fn less_safe_cbc_pkcs7( key: UnboundCipherKey, context: EncryptionContext, ) -> Result<Self, Unspecified>

Constructs a StreamingEncryptingKey for encrypting data using the CBC cipher mode with pkcs7 padding. The resulting ciphertext will be longer than the plaintext; padding is added to fill the next block of ciphertext.

This is considered less safe because the caller could potentially construct an EncryptionContext from a previously used initialization vector (IV).

Errors

Returns an error on an internal failure.