aws_lc_rs::rsa
Struct PrivateDecryptingKey

pub struct PrivateDecryptingKey(/* private fields */);

An RSA private key used for decrypting ciphertext encrypted by a PublicEncryptingKey.

Implementations

impl PrivateDecryptingKey

pub fn generate(size: KeySize) -> Result<Self, Unspecified>

Generate a new RSA private key pair for use with asymmetrical encryption.

Supports the following key sizes:

KeySize::Rsa2048

KeySize::Rsa3072

KeySize::Rsa4096

KeySize::Rsa8192

Errors

Unspecified for any error that occurs during the generation of the RSA keypair.

pub fn generate_fips(size: KeySize) -> Result<Self, Unspecified>

Generate a new RSA private key pair for use with asymmetrical encryption.

Supports the following key sizes:

KeySize::Rsa2048

KeySize::Rsa3072

KeySize::Rsa4096

Errors

Unspecified: Any key generation failure.

pub fn from_pkcs8(pkcs8: &[u8]) -> Result<Self, KeyRejected>

Construct a PrivateDecryptingKey from the provided PKCS#8 (v1) document.

Supports RSA key sizes between 2048 and 8192 (inclusive).

Errors

Unspecified for any error that occurs during deserialization of this key from PKCS#8.

pub fn is_valid_fips_key(&self) -> bool

Returns a boolean indicator if this RSA key is an approved FIPS 140-3 key.

pub fn key_size_bytes(&self) -> usize

Returns the RSA key size in bytes.

pub fn key_size_bits(&self) -> usize

Returns the RSA key size in bits.

pub fn public_key(&self) -> PublicEncryptingKey

Retrieves the PublicEncryptingKey corresponding with this PrivateDecryptingKey.