aws_lc_rs::signature
Trait VerificationAlgorithm

pub trait VerificationAlgorithm:
    Debug
    + Sync
    + Sealed {
    // Required methods
    fn verify(
        &self,
        public_key: Input<'_>,
        msg: Input<'_>,
        signature: Input<'_>,
    ) -> Result<(), Unspecified>;
    fn verify_sig(
        &self,
        public_key: &[u8],
        msg: &[u8],
        signature: &[u8],
    ) -> Result<(), Unspecified>;
}

A signature verification algorithm.

Required Methods

fn verify( &self, public_key: Input<'_>, msg: Input<'_>, signature: Input<'_>, ) -> Result<(), Unspecified>

๐Deprecated: please use VerificationAlgorithm::verify_sig instead

Verify the signature signature of message msg with the public key public_key.

Errors

error::Unspecified if inputs not verified.

fn verify_sig( &self, public_key: &[u8], msg: &[u8], signature: &[u8], ) -> Result<(), Unspecified>

Verify the signature signature of message msg with the public key public_key.

Errors

error::Unspecified if inputs not verified.