Struct aws_lc_rs::tls_prf::Secret

pub struct Secret { /* private fields */ }

Encapsulates a PRF algorithm and secret bytes to be used to derive output.

Implementations

impl Secret

pub fn new( algorithm: &'static Algorithm, secret: &[u8] ) -> Result<Self, Unspecified>

Constructs a new Secret for use with the associated P_hash Algorithm.

Errors

Unspecified: If secret.is_empty() == true.

pub fn derive( self, label: &[u8], seed: &[u8], output: usize ) -> Result<Secret, Unspecified>

Calculates len bytes of TLS PRF using the configured Algorithm, and returns Secret of length len. See RFC5246

Errors

Unspecified: Returned if the PRF derivation fails.

pub fn derive_with_seed_concatination( self, label: &[u8], seed1: &[u8], seed2: &[u8], len: usize ) -> Result<Secret, Unspecified>

Calculates len bytes of TLS PRF using the configured Algorithm, and returns Secret of length len.

In this method, seed1 and seed2 will be concatenated together: seed1 || seed2.

See RFC5246

Errors

Unspecified: Returned if the PRF derivation fails.